Restructure project with domain services folder for interfaces
Implementations that belong in infrastructure.messaging are left in pubsub.app still due to DI dependency
