+++ Kitos_PubSub.sln
@@ -30,6 +30,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PubSub.Core.DomainModel", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PubSub.Core.ApplicationServices", "PubSub.Core.ApplicationServices\PubSub.Core.ApplicationServices.csproj", "{64DE05A9-F191-42F9-9F12-11B94858278B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PubSub.Core.DomainServices", "PubSub.Core.DomainServices\PubSub.Core.DomainServices.csproj", "{BA7D647F-84B1-49BC-BE7E-E1630CEAE6EC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PubSub.Infrastructure.MessageQueue", "PubSub.Infrastructure.MessageQueue\PubSub.Infrastructure.MessageQueue.csproj", "{42BA14DB-15FB-4111-9E06-A880C1CCCAE0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -64,6 +68,14 @@ Global
 		{64DE05A9-F191-42F9-9F12-11B94858278B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{64DE05A9-F191-42F9-9F12-11B94858278B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{64DE05A9-F191-42F9-9F12-11B94858278B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BA7D647F-84B1-49BC-BE7E-E1630CEAE6EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BA7D647F-84B1-49BC-BE7E-E1630CEAE6EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BA7D647F-84B1-49BC-BE7E-E1630CEAE6EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BA7D647F-84B1-49BC-BE7E-E1630CEAE6EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42BA14DB-15FB-4111-9E06-A880C1CCCAE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42BA14DB-15FB-4111-9E06-A880C1CCCAE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42BA14DB-15FB-4111-9E06-A880C1CCCAE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42BA14DB-15FB-4111-9E06-A880C1CCCAE0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
