+++ PubSub.Application/Mapping/ISubscriptionMapper.cs
@@ -1,6 +1,6 @@
 ﻿using PubSub.Application.DTOs.Request;
 using PubSub.Application.DTOs.Response;
-using PubSub.Application.Models;
+using PubSub.Core.ApplicationServices.Models;
 using PubSub.Core.DomainModel;
 
 namespace PubSub.Application.Mapping;
