+++ PubSub.Application/Services/RabbitMQConsumerFactory.cs
@@ -1,8 +1,10 @@
-﻿using PubSub.Application.Services.RabbitMQConnection;
+﻿using PubSub.Application;
 using PubSub.Core.ApplicationServices.Notifier;
 using PubSub.Core.ApplicationServices.Serializer;
+using PubSub.Core.DomainServices.RabbitMQConnection;
+using PubSub.Core.DomainServices.Subscriber;
 
-namespace PubSub.Application.Services.Consumer
+namespace PubSub.Application.Services
 {
     public class RabbitMQConsumerFactory : IRabbitMQConsumerFactory
     {
