+++ PubSub.Infrastructure.MessageQueue/PubSub.Infrastructure.MessageQueue.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\PubSub.Application\PubSub.Application.csproj" />
+    <ProjectReference Include="..\PubSub.Core.DomainServices\PubSub.Core.DomainServices.csproj" />
+    <ProjectReference Include="..\PubSub.DataAccess\PubSub.DataAccess.csproj" />
+  </ItemGroup>
+
+</Project>
