+++ PubSub.Test/PubSub.Test.csproj
@@ -21,8 +21,12 @@
 
   <ItemGroup>
     <ProjectReference Include="..\PubSub.Application\PubSub.Application.csproj" />
+    <ProjectReference Include="..\PubSub.Core.Abstractions\PubSub.Core.Abstractions.csproj" />
     <ProjectReference Include="..\PubSub.Core.ApplicationServices\PubSub.Core.ApplicationServices.csproj" />
+    <ProjectReference Include="..\PubSub.Core.DomainModel\PubSub.Core.DomainModel.csproj" />
+    <ProjectReference Include="..\PubSub.Core.DomainServices\PubSub.Core.DomainServices.csproj" />
     <ProjectReference Include="..\PubSub.DataAccess\PubSub.DataAccess.csproj" />
+    <ProjectReference Include="..\PubSub.Infrastructure.MessageQueue\PubSub.Infrastructure.MessageQueue.csproj" />
   </ItemGroup>
 
   <ItemGroup>
